requirements.txt
flask
pyTelegramBotAPI
google-genai
supabase